Various Improvements to Acuity and other Classes
Acuity now shows the first appointment of the day is you have multiple appointments that day. Once you check-in a note is made in the appointments and that appointment is not shown again.

ClockIO and Appointment models moved into their respective parent classes as they are the owners of those methods.

Various Logging improvements through.