Replaced ID list with Problem List & Add Custom Notes
Prepare for future development. Staff at the front desk will have a live view screen and will be able to address the user's questions. Giving more power to the front desk consultant, as well as curbing favoritism.
Also, the note on the Confirmation page can be changed as needed - like to let people know about the ability to schedule ParScore appointments ahead of time.