Live Improvements (#30)
* Fixed FW-6.

* Added SSE support for Live.

Live used to poll the server for events every 2.5 seconds. This has been replaced with Server Sent Events (SSE) to simplify the process and reduce network traffic.
There are still some bugs, but those should be fixed before release.

* Improved Vault Token handeling.

Instead of refreshing the token before every use, the lease duration of the token is tracked, and is only renewed shortly before it expires.
This relies on frequent traffic, but if the token express a new one can be obtained quite easily via the AppRole Auth Backend in Vault.

* UI improvements for Live

The Login button is now disabled once it is clicked to prevent multiple logins, and to show that something is happening.

* Using Session Storage for Quotes.

Quotes have been moved from Cookies to Session Storage, and the refresh method has been updated to ensure that fresh quotes are shown when intended.