Integrate FollowUp and Alerts Modules via Quartz (#34)
* Added VR to SiteMap

* Initial Commit

* Added Testing Environment Variable to Jenkins File

* Restructured Build Steps

* Fixed Bb Auth Logging and Removed Jenkins File.

Jenkins, nice idea, pain in the butt to setup. TeamCity it is for now!

* Refactored Package Names

* Moved Alerts and FollowUp modules into Main package via Quartz Scheduler.

The Alerts and Follow Up modules are now run via the Quartz Scheduler inside of the TomCat container. This reduces the overhead of the deployment, and simplifies the code base to a degree.

* Updated Docs for new Release

Also moved CRON Scheduling String to a Vault Parameter to allow for greater scheduling flexibility.

* Removed old Shell Scripts for Alters and Follow Up modules

These are no longer needed, because this is all managed by Quartz now.