New Client UI (#39)
* Added VR to SiteMap

* Initial Commit

* Added Testing Environment Variable to Jenkins File

* Restructured Build Steps

* Fixed Bb Auth Logging and Removed Jenkins File.

Jenkins, nice idea, pain in the butt to setup. TeamCity it is for now!

* Refactored Package Names

* Moved Alerts and FollowUp modules into Main package via Quartz Scheduler.

The Alerts and Follow Up modules are now run via the Quartz Scheduler inside of the TomCat container. This reduces the overhead of the deployment, and simplifies the code base to a degree.

* Updated Docs for new Release

Also moved CRON Scheduling String to a Vault Parameter to allow for greater scheduling flexibility.

* Removed old Shell Scripts for Alters and Follow Up modules

These are no longer needed, because this is all managed by Quartz now.

* Setup jTwig and Welcome Page

* Added Type Select Page and Added Transition Animations.

* Added ID Entry Card.

* Added Goal Select and Confirmation Views

Also added backend support for Goal Select, as well as Front-End logic to support the needed interactions.

* Refactored SimpleMessage Model Class

* Added Appointment Selection and Completion

* Bug Fix

* Added Back Button to Level 2+ pages.

Level 2 pages include "Other" and "Meet with an Instructional Designer", and other pages that have a second set of buttons.

* Migrated Follow Up to jTwig.

Also updated unsubscribe script to leverage the power of jQuery!

* Client UI Improvements

Added Loading Indicator to ID Entry Card to show that something is happening. Also, some refactoring to help with later development of other components.

* Quick Clock

Allows for the quick Clocking In and Out of Staff Members.

* Restored Order to Appointment List.

Appointment list got a little out of hand with too many appointments. Appointment List has been taught the meaning of self-control (and scrolling).

* Mistakes should be forgiven.

Plus, nobody like waiting forever.

* Added Stripe Support.

Users can now use MagStripe Readers to enter their IDs. Also added Input Checking on New Endpoints.

* Removed Old Interface Assets

* Added Acuity Scheduler Support.

fixed WELCOME-24

* Added Acuity Scheduler Support.

fixed WELCOME-24

* Update Bootstrap Version.

* Removed Schedule ParScore Appointment as Option in Goal Seclect